applying auto migration, pass build
